age_groups/app/requirements.txt
fastapi
uvicorn
motor
aioredis
python-dotenv
pydantic